--- Keamanan_Apk/templates/register.html
@@ -1,33 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="d-flex justify-content-center align-items-center" style="min-height: 80vh;">
-    <div class="card shadow-lg p-4 mb-5 bg-white rounded-lg" style="max-width: 450px; width: 100%;">
-        <div class="card-body">
-            <h2 class="card-title text-center mb-4 text-primary fw-bold">Buat Akun Baru</h2>
-            <p class="card-text text-center text-secondary mb-4">Daftar sekarang untuk mulai mengelola To-Do List Anda.</p>
-
-            <form method="POST">
-                <div class="mb-3">
-                    <label for="username" class="form-label fw-bold">Username</label>
-                    <input type="text" class="form-control rounded-pill px-3 py-2" id="username" name="username" required placeholder="Pilih username Anda">
-                </div>
-                <div class="mb-3">
-                    <label for="password" class="form-label fw-bold">Password</label>
-                    <input type="password" class="form-control rounded-pill px-3 py-2" id="password" name="password" required placeholder="Buat password Anda">
-                </div>
-                <div class="mb-4">
-                    <label for="confirm_password" class="form-label fw-bold">Konfirmasi Password</label>
-                    <input type="password" class="form-control rounded-pill px-3 py-2" id="confirm_password" name="confirm_password" required placeholder="Ulangi password Anda">
-                </div>
-                <div class="d-grid gap-2">
-                    <button type="submit" class="btn btn-primary btn-lg rounded-pill fw-bold">Daftar Sekarang</button>
-                </div>
-            </form>
-
-            <div class="text-center mt-4">
-                <p class="text-muted">Sudah punya akun? <a href="{{ url_for('login') }}" class="text-primary fw-bold text-decoration-none">Login di sini</a></p>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
